# Separate affine_channel from math and optimize it