--- a/caffe2/utils/CMakeLists.txt
+++ b/caffe2/utils/CMakeLists.txt
@@ -1,17 +1,18 @@
 list(APPEND Caffe2_CPU_SRCS
-  utils/proto_convert.cc
-  utils/proto_wrap.cc
-  utils/proto_utils.cc
-  utils/murmur_hash3.cc
-  utils/smart_tensor_printer.cc
-  utils/signal_handler.cc
-  utils/string_utils.cc
-  utils/threadpool/ThreadPool.cc
-  utils/cpuid.cc
   utils/bench_utils.cc
+  utils/cpuid.cc
+  utils/math/elementwise.cc
   utils/math_cpu.cc
   utils/math_utils.cc
-  utils/thread_name.cc)
+  utils/murmur_hash3.cc
+  utils/proto_convert.cc
+  utils/proto_utils.cc
+  utils/proto_wrap.cc
+  utils/signal_handler.cc
+  utils/smart_tensor_printer.cc
+  utils/string_utils.cc
+  utils/thread_name.cc
+  utils/threadpool/ThreadPool.cc)
 
 # ---[ threadpool/pthreadpool* is a local modification of the NNPACK
 # pthreadpool with a very similar interface. Neither NNPACK, nor this
@@ -24,10 +25,12 @@ if (NOT MSVC)
 endif()
 
 set(Caffe2_GPU_SRCS ${Caffe2_GPU_SRCS}
+        utils/math/elementwise.cu
         utils/math_gpu.cu
         )
 
 set(Caffe2_HIP_SRCS ${Caffe2_HIP_SRCS}
+        utils/math/hip/elementwise.hip
         utils/hip/math_gpu.hip
         )