Summary:
Separate affine_channel from math and optimize it

i-am-not-moving-c2-to-c10

Differential Revision: D13727606