@BIT-silence 2019-01-18
Benchmark:
N = 256, C = 256, H = 128, W = 128

AffineChannelOp NCHW order: 67.47 -> 38.32ms
AffineChannelOp NHWC order: 103.95-> 39.49ms

SpatialBN NCHW testing: 38.31ms, CuDNN 43.31ms
SpatialBN NCHW training: 69.06ms, CuDNN 66.58ms